Refactor workbook navigator into modular library
