--- lib/presentation/workbook_navigator.dart
@@ -1,2707 +1,0 @@
-import 'dart:async';
-import 'package:code_text_field/code_text_field.dart';
-import 'package:flutter/foundation.dart';
-import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
-import 'package:highlight/languages/yaml.dart';
-
-import '../application/scripts/models.dart';
-import '../application/scripts/runtime.dart';
-import '../application/scripts/storage.dart';
-
-import '../application/commands/add_notes_page_command.dart';
-import '../application/commands/add_sheet_command.dart';
-import '../application/commands/command_utils.dart';
-import '../application/commands/remove_notes_page_command.dart';
-import '../application/commands/remove_sheet_command.dart';
-import '../application/commands/workbook_command_manager.dart';
-import '../domain/cell.dart';
-import '../domain/menu_page.dart';
-import '../domain/notes_page.dart';
-import '../domain/sheet.dart';
-import '../domain/workbook.dart';
-import '../domain/workbook_page.dart';
-import '../state/sheet_selection_state.dart';
-import 'themes/highlight_themes.dart';
-import 'widgets/command_ribbon.dart';
-import 'widgets/formula_bar.dart';
-import 'widgets/menu_page_view.dart';
-import 'widgets/notes_page_view.dart';
-import 'widgets/sheet_grid.dart';
-import 'widgets/workbook_page_tab_bar.dart';
-import 'widgets/top_aligned_code_field.dart';
-import 'workbook_page_display.dart';
-
-class CustomAction {
-  CustomAction({
-    required this.id,
-    required this.label,
-    required this.template,
-  });
-
-  final String id;
-  final String label;
-  final String template;
-}
-
-class _ScriptTreeNode {
-  const _ScriptTreeNode({
-    required this.id,
-    required this.label,
-    this.subtitle,
-    this.icon,
-    this.descriptor,
-    this.pageName,
-    this.rawSharedKey,
-    this.hasContent = false,
-    this.emptyLabel,
-    this.isGroup = false,
-    this.children = const <_ScriptTreeNode>[],
-  });
-
-  final String id;
-  final String label;
-  final String? subtitle;
-  final IconData? icon;
-  final ScriptDescriptor? descriptor;
-  final String? pageName;
-  final String? rawSharedKey;
-  final bool hasContent;
-  final String? emptyLabel;
-  final bool isGroup;
-  final List<_ScriptTreeNode> children;
-}
-
-class _ScriptTreeBuildResult {
-  _ScriptTreeBuildResult({
-    required this.nodes,
-    required this.parents,
-    required this.expandableIds,
-  });
-
-  final List<_ScriptTreeNode> nodes;
-  final Map<String, String?> parents;
-  final Set<String> expandableIds;
-}
-
-class ScriptEditorTab {
-  ScriptEditorTab({
-    required this.descriptor,
-    required this.controller,
-    this.pageName,
-    this.rawSharedKey,
-    this.isDirty = false,
-    this.isMutable = true,
-    this.status,
-  });
-
-  ScriptDescriptor descriptor;
-  final CodeController controller;
-  bool isDirty;
-  bool isMutable;
-  String? status;
-  String? pageName;
-  String? rawSharedKey;
-  VoidCallback? listener;
-}
-
-class WorkbookNavigator extends StatefulWidget {
-  const WorkbookNavigator({
-    super.key,
-    required this.commandManager,
-    required this.scriptRuntime,
-    required this.isAdmin,
-  });
-
-  final WorkbookCommandManager commandManager;
-  final ScriptRuntime scriptRuntime;
-  final bool isAdmin;
-
-  @override
-  State<WorkbookNavigator> createState() => _WorkbookNavigatorState();
-}
-
-class _WorkbookNavigatorState extends State<WorkbookNavigator> {
-  late PageController _pageController;
-  final Map<String, SheetSelectionState> _selectionStates =
-      <String, SheetSelectionState>{};
-  final Map<String, TextEditingController> _notesControllers =
-      <String, TextEditingController>{};
-  final Map<String, VoidCallback> _notesListeners =
-      <String, VoidCallback>{};
-  final List<CustomAction> _customActions = <CustomAction>[];
-  late final TextEditingController _customActionLabelController;
-  late final TextEditingController _customActionTemplateController;
-  final TextEditingController _sharedScriptKeyController =
-      TextEditingController(text: 'shared_module');
-  final List<ScriptEditorTab> _scriptEditorTabs = <ScriptEditorTab>[];
-  int? _activeScriptTabIndex;
-  ScriptScope _scriptEditorScope = ScriptScope.page;
-  String? _scriptEditorPageName;
-  String _scriptSharedKey = 'shared_module';
-  bool _scriptEditorLoading = false;
-  String? _scriptEditorStatus;
-  ScriptDescriptor? _currentScriptDescriptor;
-  bool _suppressScriptEditorChanges = false;
-  bool _scriptEditorFullscreen = false;
-  bool _scriptEditorSplitPreview = false;
-  bool _scriptEditorMutable = true;
-  WidgetBuilder? _scriptEditorOverlayBuilder;
-  late int _currentPageIndex;
-  final List<StoredScript> _scriptLibrary = <StoredScript>[];
-  bool _scriptLibraryLoading = false;
-  String? _scriptLibraryError;
-  final List<_ScriptTreeNode> _scriptTreeNodes = <_ScriptTreeNode>[];
-  final Map<String, bool> _scriptTreeExpanded = <String, bool>{};
-  final Map<String, String?> _scriptTreeParents = <String, String?>{};
-  final Set<String> _scriptTreeExpandableNodes = <String>{};
-  String? _activeScriptNodeId;
-
-  WorkbookCommandManager get _manager => widget.commandManager;
-  ScriptRuntime get _runtime => widget.scriptRuntime;
-  bool get _isAdmin => widget.isAdmin;
-
-  ScriptEditorTab? get _activeScriptTab {
-    final index = _activeScriptTabIndex;
-    if (index == null || index < 0 || index >= _scriptEditorTabs.length) {
-      return null;
-    }
-    return _scriptEditorTabs[index];
-  }
-
-  @override
-  void initState() {
-    super.initState();
-    _customActionLabelController = TextEditingController();
-    _customActionTemplateController = TextEditingController();
-    _sharedScriptKeyController.addListener(_handleSharedScriptKeyChanged);
-    if (_isAdmin) {
-      _initialiseCustomActions();
-      unawaited(_refreshScriptLibrary());
-    }
-    final initialPageIndex = _manager.activePageIndex;
-    final pages = _manager.workbook.pages;
-    if (pages.isNotEmpty) {
-      final safeIndex = initialPageIndex >= 0 && initialPageIndex < pages.length
-          ? initialPageIndex
-          : 0;
-      _scriptEditorPageName = pages[safeIndex].name;
-    }
-    _currentPageIndex = initialPageIndex;
-    _pageController = PageController(
-      initialPage: initialPageIndex < 0 ? 0 : initialPageIndex,
-    );
-    _manager.addListener(_handleManagerChanged);
-    _updateScriptTree(notify: false);
-    if (_isAdmin) {
-      WidgetsBinding.instance.addPostFrameCallback((_) {
-        if (mounted) {
-          unawaited(_loadScriptEditor());
-        }
-      });
-    }
-  }
-
-  @override
-  void didUpdateWidget(covariant WorkbookNavigator oldWidget) {
-    super.didUpdateWidget(oldWidget);
-    if (oldWidget.commandManager != widget.commandManager) {
-      oldWidget.commandManager.removeListener(_handleManagerChanged);
-      final newIndex = widget.commandManager.activePageIndex;
-      _currentPageIndex = newIndex;
-      _pageController.dispose();
-      _pageController = PageController(
-        initialPage: newIndex < 0 ? 0 : newIndex,
-      );
-      widget.commandManager.addListener(_handleManagerChanged);
-    }
-    if (!oldWidget.isAdmin && widget.isAdmin) {
-      _initialiseCustomActions();
-      WidgetsBinding.instance.addPostFrameCallback((_) {
-        if (!mounted) {
-          return;
-        }
-        unawaited(_refreshScriptLibrary());
-        unawaited(_loadScriptEditor());
-      });
-      _updateScriptTree();
-    }
-    if (oldWidget.isAdmin && !widget.isAdmin) {
-      _updateScriptTree();
-    }
-  }
-
-  void _handleManagerChanged() {
-    final workbook = _manager.workbook;
-    final sheets = workbook.sheets;
-    final notesPages = workbook.pages.whereType<NotesPage>().toList();
-
-    final removedSheets = _selectionStates.keys
-        .where((name) => sheets.every((sheet) => sheet.name != name))
-        .toList(growable: false);
-    for (final sheet in removedSheets) {
-      _selectionStates.remove(sheet)?.dispose();
-    }
-
-    final removedNotes = _notesControllers.keys
-        .where((name) => notesPages.every((page) => page.name != name))
-        .toList(growable: false);
-    for (final noteName in removedNotes) {
-      final controller = _notesControllers.remove(noteName);
-      final listener = _notesListeners.remove(noteName);
-      if (controller != null && listener != null) {
-        controller.removeListener(listener);
-      }
-      controller?.dispose();
-    }
-
-    final newIndex = _manager.activePageIndex;
-    if (newIndex != _currentPageIndex) {
-      _commitEditsForPage(workbook, _currentPageIndex);
-      _currentPageIndex = newIndex;
-      _jumpToPage(newIndex);
-    }
-    if (_scriptEditorScope == ScriptScope.page) {
-      final pageNames = workbook.pages.map((page) => page.name).toList();
-      final currentSelection = _scriptEditorPageName;
-      if (currentSelection == null || !pageNames.contains(currentSelection)) {
-        final fallback = pageNames.isNotEmpty ? pageNames.first : null;
-        if (fallback != currentSelection) {
-          setState(() {
-            _scriptEditorPageName = fallback;
-          });
-          if (_isAdmin && fallback != null) {
-            unawaited(_loadScriptEditor());
-          }
-        }
-      }
-    }
-    if (_isAdmin) {
-      unawaited(_synchronisePageScriptsWithWorkbook(workbook));
-      _updateScriptTree();
-    }
-  }
-
-  Future<void> _synchronisePageScriptsWithWorkbook(Workbook workbook) async {
-    if (!_isAdmin) {
-      return;
-    }
-    final createdPages = <String>[];
-    for (final page in workbook.pages) {
-      final descriptor = ScriptDescriptor(
-        scope: ScriptScope.page,
-        key: normaliseScriptKey(page.name),
-      );
-      if (_hasScriptDescriptor(descriptor)) {
-        continue;
-      }
-      final existing = await _runtime.storage.loadScript(descriptor);
-      if (existing != null) {
-        continue;
-      }
-      final template = _defaultScriptTemplate(descriptor);
-      final stored = await _runtime.storage.saveScript(descriptor, template);
-      createdPages.add(page.name);
-      debugPrint(
-        'Script automatiquement créé pour la page "${page.name}" (${stored.origin}).',
-      );
-    }
-    if (createdPages.isNotEmpty && mounted) {
-      setState(() {
-        final notice =
-            'Scripts créés automatiquement pour: ${createdPages.join(', ')}.';
-        _scriptEditorStatus =
-            _scriptEditorStatus == null ? notice : '${_scriptEditorStatus!}\n$notice';
-      });
-    }
-    await _refreshScriptLibrary(silent: true);
-  }
-
-  void _jumpToPage(int index) {
-    if (index < 0) {
-      return;
-    }
-    final pageCount = _manager.workbook.pages.length;
-    if (pageCount == 0) {
-      return;
-    }
-    final targetIndex = index >= pageCount ? pageCount - 1 : index;
-    if (!_pageController.hasClients) {
-      WidgetsBinding.instance.addPostFrameCallback((_) {
-        if (mounted) {
-          _pageController.jumpToPage(targetIndex);
-        }
-      });
-      return;
-    }
-    if (_pageController.page?.round() == targetIndex) {
-      return;
-    }
-    _pageController.animateToPage(
-      targetIndex,
-      duration: const Duration(milliseconds: 200),
-      curve: Curves.easeInOut,
-    );
-  }
-
-  @override
-  void dispose() {
-    _manager.removeListener(_handleManagerChanged);
-    _sharedScriptKeyController.removeListener(_handleSharedScriptKeyChanged);
-    _pageController.dispose();
-    for (final state in _selectionStates.values) {
-      state.dispose();
-    }
-    _notesControllers.forEach((name, controller) {
-      final listener = _notesListeners[name];
-      if (listener != null) {
-        controller.removeListener(listener);
-      }
-      controller.dispose();
-    });
-    _customActionLabelController.dispose();
-    _customActionTemplateController.dispose();
-    for (final tab in _scriptEditorTabs) {
-      final listener = tab.listener;
-      if (listener != null) {
-        tab.controller.removeListener(listener);
-      }
-      tab.controller.dispose();
-    }
-    _sharedScriptKeyController.dispose();
-    super.dispose();
-  }
-
-  SheetSelectionState _stateForSheet(Workbook workbook, Sheet sheet) {
-    final state = _selectionStates.putIfAbsent(
-      sheet.name,
-      () => SheetSelectionState(
-        onValuesChanged: (values) => _persistSheetValues(sheet.name, values),
-      ),
-    );
-    state.onValuesChanged =
-        (values) => _persistSheetValues(sheet.name, values);
-    state.syncFromSheet(sheet);
-    return state;
-  }
-
-  void _commitEditsForSheet(String sheetName) {
-    final state = _selectionStates[sheetName];
-    state?.commitEditingValue();
-  }
-
-  void _commitEditsForPage(Workbook workbook, int pageIndex) {
-    if (pageIndex < 0 || pageIndex >= workbook.pages.length) {
-      return;
-    }
-    final page = workbook.pages[pageIndex];
-    if (page is Sheet) {
-      _commitEditsForSheet(page.name);
-    }
-  }
-
-  TextEditingController _controllerForNotesPage(NotesPage page) {
-    final existing = _notesControllers[page.name];
-    if (existing != null) {
-      if (existing.text != page.content) {
-        final listener = _notesListeners[page.name];
-        if (listener != null) {
-          existing.removeListener(listener);
-        }
-        final previousSelection = existing.selection;
-        existing.value = existing.value.copyWith(
-          text: page.content,
-          selection: previousSelection,
-          composing: TextRange.empty,
-        );
-        if (listener != null) {
-          existing.addListener(listener);
-        }
-      }
-      return existing;
-    }
-    final controller = TextEditingController(text: page.content);
-    void handleChange() => _handleNotesChanged(page.name, controller.text);
-    controller.addListener(handleChange);
-    _notesControllers[page.name] = controller;
-    _notesListeners[page.name] = handleChange;
-    return controller;
-  }
-
-  void _handleNotesChanged(String pageName, String content) {
-
-    final workbook = _manager.workbook;
-
-    final pageIndex = workbook.pages.indexWhere(
-
-      (page) => page.name == pageName,
-
-    );
-
-    if (pageIndex == -1) {
-
-      return;
-
-    }
-
-    final page = workbook.pages[pageIndex];
-
-    if (page is! NotesPage) {
-
-      return;
-
-    }
-
-    if (page.content == content) {
-
-      return;
-
-    }
-
-    final updatedPage = page.copyWith(content: content);
-
-    final updatedWorkbook = replacePage(workbook, pageIndex, updatedPage);
-
-    _manager.applyExternalUpdate(
-
-      updatedWorkbook,
-
-      activePageIndex: _manager.activePageIndex,
-
-    );
-
-    unawaited(
-
-      _runtime.dispatchNotesChanged(page: updatedPage, content: content),
-
-    );
-
-  }
-
-
-
-  void _handleSelectPage(int pageIndex) {
-    final workbook = _manager.workbook;
-    if (pageIndex < 0 || pageIndex >= workbook.pages.length) {
-      return;
-    }
-    _commitEditsForPage(workbook, _currentPageIndex);
-    _manager.setActivePage(pageIndex);
-  }
-
-  void _handleAddSheet() {
-    _commitActiveSelectionEdits();
-    _manager.execute(AddSheetCommand());
-  }
-
-  void _handleAddNotesPage() {
-    _commitActiveSelectionEdits();
-    _manager.execute(AddNotesPageCommand());
-  }
-
-  void _handleRemovePage(int pageIndex) {
-    _commitActiveSelectionEdits();
-    final workbook = _manager.workbook;
-    if (pageIndex < 0 || pageIndex >= workbook.pages.length) {
-      return;
-    }
-    final page = workbook.pages[pageIndex];
-    if (page is Sheet) {
-      final sheetIndex = workbook.sheets.indexOf(page);
-      if (sheetIndex != -1) {
-        _selectionStates.remove(page.name)?.dispose();
-        _manager.execute(RemoveSheetCommand(sheetIndex: sheetIndex));
-      }
-      return;
-    }
-    if (page is NotesPage) {
-      _manager.execute(RemoveNotesPageCommand(pageIndex: pageIndex));
-    }
-  }
-
-  bool _canRemovePage(Workbook workbook, int pageIndex) {
-    if (pageIndex < 0 || pageIndex >= workbook.pages.length) {
-      return false;
-    }
-    final page = workbook.pages[pageIndex];
-    if (page is MenuPage) {
-      return false;
-    }
-    if (page is Sheet) {
-      return workbook.sheets.length > 1;
-    }
-    return true;
-  }
-
-  void _handleRemoveSheet(int sheetIndex) {
-    _commitActiveSelectionEdits();
-    final workbook = _manager.workbook;
-    if (sheetIndex < 0 || sheetIndex >= workbook.sheets.length) {
-      return;
-    }
-    final sheetName = workbook.sheets[sheetIndex].name;
-    _selectionStates.remove(sheetName)?.dispose();
-    _manager.execute(RemoveSheetCommand(sheetIndex: sheetIndex));
-  }
-
-  void _commitActiveSelectionEdits() {
-    _commitEditsForPage(_manager.workbook, _manager.activePageIndex);
-  }
-
-  void _persistSheetValues(
-    String sheetName,
-    Map<CellPosition, String> values,
-  ) {
-    final workbook = _manager.workbook;
-    Sheet? targetSheet;
-    for (final sheet in workbook.sheets) {
-      if (sheet.name == sheetName) {
-        targetSheet = sheet;
-        break;
-      }
-    }
-    if (targetSheet == null) {
-      return;
-    }
-
-    final pageIndex = workbook.pages.indexOf(targetSheet);
-    if (pageIndex == -1) {
-      return;
-    }
-
-    final rows = cloneSheetRows(targetSheet);
-    var didChange = false;
-    for (var r = 0; r < targetSheet.rowCount; r++) {
-      final row = rows[r];
-      for (var c = 0; c < targetSheet.columnCount; c++) {
-        final position = CellPosition(r, c);
-        final text = values[position];
-        if (text == null || text.isEmpty) {
-          final cell = row[c];
-          if (cell.type != CellType.empty || cell.value != null) {
-            row[c] = Cell(row: r, column: c, type: CellType.empty, value: null);
-            didChange = true;
-          }
-          continue;
-        }
-
-        final nextCell = Cell.fromValue(row: r, column: c, value: text);
-        final currentCell = row[c];
-        if (currentCell.type != nextCell.type ||
-            currentCell.value != nextCell.value) {
-          row[c] = nextCell;
-          didChange = true;
-        }
-      }
-    }
-
-    if (!didChange) {
-      return;
-    }
-
-    final updatedSheet = rebuildSheetFromRows(targetSheet, rows);
-    final updatedWorkbook = replaceSheetAtPageIndex(
-      workbook,
-      pageIndex,
-      updatedSheet,
-    );
-    _manager.applyExternalUpdate(
-      updatedWorkbook,
-      activePageIndex: _manager.activePageIndex,
-    );
-  }
-
-  void _handleScriptEditorChanged(ScriptEditorTab tab) {
-    if (_suppressScriptEditorChanges) {
-      return;
-    }
-    if (!tab.isMutable) {
-      return;
-    }
-    final isActive = identical(tab, _activeScriptTab);
-    if (!tab.isDirty || (isActive && _scriptEditorSplitPreview)) {
-      setState(() {
-        tab.isDirty = true;
-      });
-    } else {
-      tab.isDirty = true;
-    }
-  }
-
-  void _handleSharedScriptKeyChanged() {
-    final value = normaliseScriptKey(_sharedScriptKeyController.text);
-    if (value != _scriptSharedKey) {
-      setState(() {
-        _scriptSharedKey = value;
-      });
-    }
-    final activeTab = _activeScriptTab;
-    if (activeTab != null &&
-        activeTab.descriptor.scope == ScriptScope.shared) {
-      activeTab.rawSharedKey = _sharedScriptKeyController.text;
-    }
-  }
-
-  Future<void> _refreshScriptLibrary({bool silent = false}) async {
-    if (!_isAdmin) {
-      return;
-    }
-    if (!silent) {
-      setState(() {
-        _scriptLibraryLoading = true;
-        _scriptLibraryError = null;
-      });
-    }
-    try {
-      final scripts = await _runtime.storage.loadAll();
-      if (!mounted) {
-        return;
-      }
-      setState(() {
-        _scriptLibrary
-          ..clear()
-          ..addAll(scripts);
-        _scriptLibraryLoading = false;
-        _scriptLibraryError = null;
-      });
-      _updateScriptTree();
-    } catch (error) {
-      if (!mounted) {
-        return;
-      }
-      setState(() {
-        _scriptLibraryLoading = false;
-        _scriptLibraryError = 'Erreur de chargement: $error';
-      });
-      _updateScriptTree();
-    }
-  }
-
-  void _updateScriptTree({bool notify = true}) {
-    if (!_isAdmin) {
-      void clearData() {
-        _scriptTreeNodes.clear();
-        _scriptTreeParents.clear();
-        _scriptTreeExpandableNodes.clear();
-        _scriptTreeExpanded.clear();
-        _activeScriptNodeId = null;
-      }
-
-      if (notify) {
-        setState(clearData);
-      } else {
-        clearData();
-      }
-      return;
-    }
-
-    final result = _computeScriptTree();
-    final descriptor = _descriptorForSelection();
-    String? activeNodeId = _activeScriptNodeId;
-    final newExpanded = <String, bool>{};
-    for (final id in result.expandableIds) {
-      if (_scriptTreeExpanded.containsKey(id)) {
-        newExpanded[id] = _scriptTreeExpanded[id]!;
-      } else {
-        newExpanded[id] = true;
-      }
-    }
-
-    if (descriptor != null) {
-      final matchedId = _findNodeIdForDescriptor(result.nodes, descriptor);
-      activeNodeId = matchedId;
-      if (matchedId != null) {
-        _expandAncestorsForId(
-          matchedId,
-          newExpanded,
-          parents: result.parents,
-        );
-      }
-    } else {
-      activeNodeId = null;
-    }
-
-    void apply() {
-      _scriptTreeNodes
-        ..clear()
-        ..addAll(result.nodes);
-      _scriptTreeParents
-        ..clear()
-        ..addAll(result.parents);
-      _scriptTreeExpandableNodes
-        ..clear()
-        ..addAll(result.expandableIds);
-      _scriptTreeExpanded
-        ..clear()
-        ..addAll(newExpanded);
-      _activeScriptNodeId = activeNodeId;
-    }
-
-    if (notify) {
-      setState(apply);
-    } else {
-      apply();
-    }
-  }
-
-  _ScriptTreeBuildResult _computeScriptTree() {
-    final workbook = _manager.workbook;
-    final pages = workbook.pages;
-    final sharedScripts = _scriptLibrary
-        .where((script) => script.descriptor.scope == ScriptScope.shared)
-        .toList()
-      ..sort((a, b) => a.descriptor.key.compareTo(b.descriptor.key));
-
-    final globalDescriptor =
-        const ScriptDescriptor(scope: ScriptScope.global, key: 'default');
-    final globalHasScript = _hasScriptDescriptor(globalDescriptor);
-
-    final nodes = <_ScriptTreeNode>[
-      _ScriptTreeNode(
-        id: 'group:workbook',
-        label: 'Classeur',
-        isGroup: true,
-        children: [
-          _ScriptTreeNode(
-            id: 'script:global',
-            label: 'Script global',
-            subtitle: globalHasScript
-                ? 'Script existant'
-                : 'Déclenché pour tout le classeur',
-            icon: Icons.language,
-            descriptor: globalDescriptor,
-            hasContent: globalHasScript,
-          ),
-        ],
-      ),
-      _ScriptTreeNode(
-        id: 'group:pages',
-        label: 'Pages',
-        isGroup: true,
-        emptyLabel: 'Aucune page disponible.',
-        children: pages
-            .map((page) {
-              final descriptor = ScriptDescriptor(
-                scope: ScriptScope.page,
-                key: normaliseScriptKey(page.name),
-              );
-              final hasScript = _hasScriptDescriptor(descriptor);
-              return _ScriptTreeNode(
-                id: 'page:${descriptor.key}',
-                label: page.name,
-                subtitle: hasScript
-                    ? 'Script existant'
-                    : 'Créer un script pour cette page',
-                icon: Icons.grid_on_outlined,
-                descriptor: descriptor,
-                pageName: page.name,
-                hasContent: hasScript,
-              );
-            })
-            .toList(growable: false),
-      ),
-      _ScriptTreeNode(
-        id: 'group:shared',
-        label: 'Modules partagés',
-        isGroup: true,
-        emptyLabel: 'Créez un module pour factoriser vos snippets.',
-        children: sharedScripts
-            .map((script) {
-              final descriptor = script.descriptor;
-              return _ScriptTreeNode(
-                id: 'shared:${descriptor.key}',
-                label: descriptor.key,
-                subtitle: 'Module partagé',
-                icon: Icons.extension,
-                descriptor: descriptor,
-                rawSharedKey: descriptor.key,
-                hasContent: true,
-              );
-            })
-            .toList(growable: false),
-      ),
-    ];
-
-    final parents = <String, String?>{};
-    final expandableIds = <String>{};
-
-    void registerNodes(List<_ScriptTreeNode> list, {String? parent}) {
-      for (final node in list) {
-        parents[node.id] = parent;
-        if (node.isGroup) {
-          expandableIds.add(node.id);
-        }
-        if (node.children.isNotEmpty) {
-          registerNodes(node.children, parent: node.id);
-        }
-      }
-    }
-
-    registerNodes(nodes);
-
-    return _ScriptTreeBuildResult(
-      nodes: nodes,
-      parents: parents,
-      expandableIds: expandableIds,
-    );
-  }
-
-  void _expandAncestorsForId(
-    String nodeId,
-    Map<String, bool> expanded, {
-    Map<String, String?>? parents,
-  }) {
-    final map = parents ?? _scriptTreeParents;
-    var current = map[nodeId];
-    while (current != null) {
-      expanded[current] = true;
-      current = map[current];
-    }
-  }
-
-  String? _findNodeIdForDescriptor(
-    List<_ScriptTreeNode> nodes,
-    ScriptDescriptor descriptor,
-  ) {
-    for (final node in nodes) {
-      final nodeDescriptor = node.descriptor;
-      if (nodeDescriptor != null &&
-          nodeDescriptor.scope == descriptor.scope &&
-          nodeDescriptor.key == descriptor.key) {
-        return node.id;
-      }
-      final childResult = _findNodeIdForDescriptor(node.children, descriptor);
-      if (childResult != null) {
-        return childResult;
-      }
-    }
-    return null;
-  }
-
-  void _toggleScriptTreeExpansion() {
-    if (_scriptTreeExpandableNodes.isEmpty) {
-      return;
-    }
-    final hasCollapsed = _scriptTreeExpandableNodes
-        .any((id) => !(_scriptTreeExpanded[id] ?? true));
-    final target = hasCollapsed;
-    setState(() {
-      for (final id in _scriptTreeExpandableNodes) {
-        _scriptTreeExpanded[id] = target;
-      }
-      if (!target && _activeScriptNodeId != null) {
-        _expandAncestorsForId(_activeScriptNodeId!, _scriptTreeExpanded);
-      }
-    });
-  }
-
-  void _handleToggleScriptGroup(String nodeId) {
-    final current = _scriptTreeExpanded[nodeId] ?? true;
-    setState(() {
-      _scriptTreeExpanded[nodeId] = !current;
-    });
-  }
-
-  Widget _buildScriptTreeNode(
-    BuildContext context,
-    _ScriptTreeNode node, {
-    int depth = 0,
-  }) {
-    final theme = Theme.of(context);
-    final padding = EdgeInsets.only(left: 16.0 * depth + 16, right: 16);
-
-    if (node.isGroup) {
-      final expanded = _scriptTreeExpanded[node.id] ?? true;
-      final children = node.children.isNotEmpty
-          ? node.children
-              .map(
-                (child) => _buildScriptTreeNode(
-                  context,
-                  child,
-                  depth: depth + 1,
-                ),
-              )
-              .toList(growable: false)
-          : <Widget>[];
-      final placeholder = node.emptyLabel;
-
-      return Column(
-        crossAxisAlignment: CrossAxisAlignment.stretch,
-        children: [
-          ListTile(
-            dense: true,
-            visualDensity: VisualDensity.compact,
-            contentPadding: padding,
-            title: Text(
-              node.label,
-              style: theme.textTheme.titleSmall?.copyWith(
-                fontWeight: FontWeight.w600,
-              ),
-            ),
-            trailing: Icon(expanded ? Icons.expand_less : Icons.expand_more),
-            onTap: () => _handleToggleScriptGroup(node.id),
-          ),
-          ClipRect(
-            child: AnimatedSize(
-              duration: const Duration(milliseconds: 200),
-              curve: Curves.easeInOut,
-              alignment: Alignment.topCenter,
-              child: expanded
-                  ? Column(
-                      crossAxisAlignment: CrossAxisAlignment.stretch,
-                      children: [
-                        if (children.isNotEmpty) ...children,
-                        if (children.isEmpty && placeholder != null)
-                          Padding(
-                            padding: EdgeInsets.fromLTRB(
-                              16.0 * (depth + 1) + 16,
-                              4,
-                              16,
-                              12,
-                            ),
-                            child: Text(
-                              placeholder,
-                              style: theme.textTheme.bodySmall,
-                            ),
-                          ),
-                      ],
-                    )
-                  : const SizedBox.shrink(),
-            ),
-          ),
-        ],
-      );
-    }
-
-    final isSelected = _activeScriptNodeId == node.id;
-
-    return ListTile(
-      dense: true,
-      visualDensity: VisualDensity.compact,
-      contentPadding: padding,
-      leading: node.icon != null ? Icon(node.icon, size: 20) : null,
-      title: Text(node.label),
-      subtitle: node.subtitle != null
-          ? Text(
-              node.subtitle!,
-              style: theme.textTheme.bodySmall,
-            )
-          : null,
-      trailing: node.hasContent
-          ? Icon(
-              Icons.check_circle,
-              size: 16,
-              color: theme.colorScheme.primary,
-            )
-          : null,
-      selected: isSelected,
-      selectedTileColor: theme.colorScheme.primary.withOpacity(0.08),
-      shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(8)),
-      onTap: node.descriptor == null
-          ? null
-          : () => _handleSelectScriptDescriptor(
-                node.descriptor!,
-                pageName: node.pageName,
-                rawSharedKey: node.rawSharedKey,
-                nodeId: node.id,
-              ),
-    );
-  }
-
-  ScriptDescriptor? _descriptorForSelection() {
-    switch (_scriptEditorScope) {
-      case ScriptScope.global:
-        return const ScriptDescriptor(scope: ScriptScope.global, key: 'default');
-      case ScriptScope.page:
-        final pageName = _scriptEditorPageName;
-        if (pageName == null || pageName.isEmpty) {
-          return null;
-        }
-        return ScriptDescriptor(
-          scope: ScriptScope.page,
-          key: normaliseScriptKey(pageName),
-        );
-      case ScriptScope.shared:
-        if (_scriptSharedKey.isEmpty) {
-          return null;
-        }
-        return ScriptDescriptor(scope: ScriptScope.shared, key: _scriptSharedKey);
-    }
-  }
-
-  bool _hasScriptDescriptor(ScriptDescriptor descriptor) {
-    return _scriptLibrary.any(
-      (script) =>
-          script.descriptor.scope == descriptor.scope &&
-          script.descriptor.key == descriptor.key,
-    );
-  }
-
-  int _indexOfTabForDescriptor(ScriptDescriptor descriptor) {
-    return _scriptEditorTabs.indexWhere(
-      (tab) =>
-          tab.descriptor.scope == descriptor.scope &&
-          tab.descriptor.key == descriptor.key,
-    );
-  }
-
-  void _activateTab(
-    int index, {
-    Map<String, bool>? expandedOverride,
-    String? resolvedNodeId,
-    String? pageNameOverride,
-    String? rawSharedKeyOverride,
-  }) {
-    if (index < 0 || index >= _scriptEditorTabs.length) {
-      setState(() {
-        _activeScriptTabIndex = null;
-        _currentScriptDescriptor = null;
-        _scriptEditorMutable = false;
-        _scriptEditorStatus =
-            'Selectionnez un script a charger pour commencer.';
-        _activeScriptNodeId = null;
-      });
-      return;
-    }
-
-    final tab = _scriptEditorTabs[index];
-    final descriptor = tab.descriptor;
-    final updatedExpanded = Map<String, bool>.from(
-      expandedOverride ?? _scriptTreeExpanded,
-    );
-    final parents = _scriptTreeParents;
-    final nodeId = resolvedNodeId ??
-        _findNodeIdForDescriptor(_scriptTreeNodes, descriptor);
-    if (nodeId != null) {
-      _expandAncestorsForId(nodeId, updatedExpanded, parents: parents);
-    }
-
-    String? resolvedPageName = pageNameOverride ?? tab.pageName;
-    if (descriptor.scope == ScriptScope.page && resolvedPageName == null) {
-      for (final page in _manager.workbook.pages) {
-        if (normaliseScriptKey(page.name) == descriptor.key) {
-          resolvedPageName = page.name;
-          break;
-        }
-      }
-    }
-    tab.pageName = resolvedPageName;
-
-    String? resolvedRawSharedKey = rawSharedKeyOverride ?? tab.rawSharedKey;
-    if (descriptor.scope == ScriptScope.shared) {
-      resolvedRawSharedKey ??= descriptor.key;
-      tab.rawSharedKey = resolvedRawSharedKey;
-    }
-
-    setState(() {
-      _activeScriptTabIndex = index;
-      _currentScriptDescriptor = descriptor;
-      _scriptEditorScope = descriptor.scope;
-      switch (descriptor.scope) {
-        case ScriptScope.global:
-          break;
-        case ScriptScope.page:
-          _scriptEditorPageName = resolvedPageName;
-          break;
-        case ScriptScope.shared:
-          final rawValue = resolvedRawSharedKey ?? descriptor.key;
-          _sharedScriptKeyController
-              .removeListener(_handleSharedScriptKeyChanged);
-          _sharedScriptKeyController.text = rawValue;
-          _sharedScriptKeyController.selection =
-              TextSelection.collapsed(offset: rawValue.length);
-          _sharedScriptKeyController
-              .addListener(_handleSharedScriptKeyChanged);
-          _scriptSharedKey = normaliseScriptKey(rawValue);
-          break;
-      }
-      _scriptTreeExpanded
-        ..clear()
-        ..addAll(updatedExpanded);
-      _activeScriptNodeId = nodeId;
-      _scriptEditorMutable = tab.isMutable;
-      _scriptEditorStatus = tab.status;
-    });
-  }
-
-  Future<void> _openOrFocusTab(
-    ScriptDescriptor descriptor, {
-    String? pageName,
-    String? rawSharedKey,
-    String? nodeId,
-    Map<String, bool>? expanded,
-  }) async {
-    final resolvedRaw =
-        descriptor.scope == ScriptScope.shared ? (rawSharedKey ?? descriptor.key) : null;
-    final existingIndex = _indexOfTabForDescriptor(descriptor);
-    if (existingIndex != -1) {
-      final tab = _scriptEditorTabs[existingIndex];
-      if (pageName != null) {
-        tab.pageName = pageName;
-      }
-      if (resolvedRaw != null) {
-        tab.rawSharedKey = resolvedRaw;
-      }
-      _activateTab(
-        existingIndex,
-        expandedOverride: expanded,
-        resolvedNodeId: nodeId,
-        pageNameOverride: pageName,
-        rawSharedKeyOverride: resolvedRaw,
-      );
-      return;
-    }
-
-    final controller = CodeController(
-      language: yaml,
-      params: const EditorParams(tabSpaces: 2),
-    );
-    final tab = ScriptEditorTab(
-      descriptor: descriptor,
-      controller: controller,
-      pageName: pageName,
-      rawSharedKey: resolvedRaw,
-      isDirty: false,
-      isMutable: false,
-    );
-    final listener = () => _handleScriptEditorChanged(tab);
-    tab.listener = listener;
-    controller.addListener(listener);
-    setState(() {
-      _scriptEditorTabs.add(tab);
-    });
-    final newIndex = _scriptEditorTabs.length - 1;
-    _activateTab(
-      newIndex,
-      expandedOverride: expanded,
-      resolvedNodeId: nodeId,
-      pageNameOverride: pageName,
-      rawSharedKeyOverride: resolvedRaw,
-    );
-    await _loadTabSource(tab);
-  }
-
-  Future<void> _loadTabSource(ScriptEditorTab tab) async {
-    setState(() {
-      _scriptEditorLoading = true;
-      if (identical(tab, _activeScriptTab)) {
-        _scriptEditorStatus = 'Chargement de ${tab.descriptor.fileName}...';
-      }
-    });
-    try {
-      final storage = _runtime.storage;
-      final existing = await storage.loadScript(tab.descriptor);
-      StoredScript? stored = existing;
-      var createdFromTemplate = false;
-      if (stored == null && storage.supportsFileSystem) {
-        final template = _defaultScriptTemplate(tab.descriptor);
-        stored = await storage.saveScript(tab.descriptor, template);
-        createdFromTemplate = true;
-        debugPrint(
-          'Script manquant pour ${tab.descriptor.fileName}. Modèle sauvegardé dans ${stored.origin}.',
-        );
-        await _refreshScriptLibrary(silent: true);
-      }
-      if (!mounted) {
-        return;
-      }
-      setState(() {
-        _currentScriptDescriptor = tab.descriptor;
-        _suppressScriptEditorChanges = true;
-        tab.controller.text = stored?.source ?? '';
-        _suppressScriptEditorChanges = false;
-        tab.isDirty = false;
-        tab.isMutable = stored?.isMutable ?? false;
-        if (stored == null) {
-          tab.status = storage.supportsFileSystem
-              ? 'Script introuvable pour ${tab.descriptor.fileName}.'
-              :
-                  'Script introuvable et édition indisponible sur cette plateforme (lecture seule).';
-        } else if (!tab.isMutable) {
-          tab.status =
-              'Script chargé depuis ${stored.origin}. Edition indisponible sur cette plateforme (lecture seule).';
-        } else {
-          tab.status = createdFromTemplate
-              ? 'Script absent. Modèle par défaut créé et sauvegardé (${stored.origin}).'
-              : 'Script chargé depuis ${stored.origin}.';
-        }
-        if (identical(tab, _activeScriptTab)) {
-          _scriptEditorMutable = tab.isMutable;
-          _scriptEditorStatus = tab.status;
-        }
-      });
-    } catch (error) {
-      if (!mounted) {
-        return;
-      }
-      setState(() {
-        final message = 'Erreur lors du chargement: $error';
-        tab.status = message;
-        if (identical(tab, _activeScriptTab)) {
-          _scriptEditorStatus = message;
-        }
-      });
-    } finally {
-      if (!mounted) {
-        return;
-      }
-      setState(() {
-        _scriptEditorLoading = false;
-      });
-    }
-  }
-
-  String _tabTitle(ScriptEditorTab tab) {
-    switch (tab.descriptor.scope) {
-      case ScriptScope.global:
-        return 'Global';
-      case ScriptScope.page:
-        return tab.pageName ?? tab.descriptor.key;
-      case ScriptScope.shared:
-        return tab.rawSharedKey ?? tab.descriptor.key;
-    }
-  }
-
-  void _handleCloseScriptTab(int index) {
-    if (index < 0 || index >= _scriptEditorTabs.length) {
-      return;
-    }
-    final tab = _scriptEditorTabs[index];
-    final listener = tab.listener;
-    if (listener != null) {
-      tab.controller.removeListener(listener);
-    }
-    tab.controller.dispose();
-
-    int? nextActiveIndex = _activeScriptTabIndex;
-    if (nextActiveIndex != null) {
-      if (nextActiveIndex == index) {
-        nextActiveIndex = null;
-      } else if (nextActiveIndex > index) {
-        nextActiveIndex -= 1;
-      }
-    }
-
-    setState(() {
-      _scriptEditorTabs.removeAt(index);
-      _activeScriptTabIndex = nextActiveIndex;
-    });
-
-    if (_scriptEditorTabs.isEmpty) {
-      setState(() {
-        _currentScriptDescriptor = null;
-        _scriptEditorMutable = false;
-        _scriptEditorStatus =
-            'Selectionnez un script a charger pour commencer.';
-        _activeScriptNodeId = null;
-      });
-      return;
-    }
-
-    final targetIndex = nextActiveIndex ?? (index > 0 ? index - 1 : 0);
-    _activateTab(targetIndex);
-  }
-
-  Widget _buildScriptTabChip(
-    BuildContext context, {
-    required ScriptEditorTab tab,
-    required bool isActive,
-    required VoidCallback onSelect,
-    required VoidCallback onClose,
-  }) {
-    final theme = Theme.of(context);
-    final colorScheme = theme.colorScheme;
-    final title = _tabTitle(tab);
-    final displayTitle = tab.isDirty ? '$title*' : title;
-    final backgroundColor = isActive
-        ? colorScheme.primary.withOpacity(0.12)
-        : colorScheme.surfaceVariant.withOpacity(
-            theme.brightness == Brightness.dark ? 0.35 : 0.25,
-          );
-    final textStyle = theme.textTheme.bodyMedium?.copyWith(
-      color: isActive ? colorScheme.primary : null,
-      fontWeight: isActive ? FontWeight.w600 : FontWeight.w500,
-    );
-
-    return Material(
-      color: backgroundColor,
-      borderRadius: BorderRadius.circular(8),
-      child: InkWell(
-        borderRadius: BorderRadius.circular(8),
-        onTap: onSelect,
-        child: Row(
-          mainAxisSize: MainAxisSize.min,
-          children: [
-            Padding(
-              padding: const EdgeInsets.symmetric(horizontal: 12, vertical: 8),
-              child: Text(displayTitle, style: textStyle),
-            ),
-            IconButton(
-              icon: const Icon(Icons.close, size: 16),
-              tooltip: 'Fermer $title',
-              padding: const EdgeInsets.all(8),
-              constraints: const BoxConstraints(minHeight: 32, minWidth: 32),
-              splashRadius: 18,
-              onPressed: onClose,
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-
-  Future<void> _handleSelectScriptDescriptor(
-    ScriptDescriptor descriptor, {
-    String? pageName,
-    String? rawSharedKey,
-    String? nodeId,
-  }) async {
-    final resolvedNodeId =
-        nodeId ?? _findNodeIdForDescriptor(_scriptTreeNodes, descriptor);
-    final expanded = Map<String, bool>.from(_scriptTreeExpanded);
-    if (resolvedNodeId != null) {
-      _expandAncestorsForId(resolvedNodeId, expanded);
-    }
-    String? resolvedPageName = pageName;
-    if (descriptor.scope == ScriptScope.page && resolvedPageName == null) {
-      for (final page in _manager.workbook.pages) {
-        if (normaliseScriptKey(page.name) == descriptor.key) {
-          resolvedPageName = page.name;
-          break;
-        }
-      }
-    }
-    final resolvedRawSharedKey = descriptor.scope == ScriptScope.shared
-        ? (rawSharedKey ?? descriptor.key)
-        : null;
-    await _openOrFocusTab(
-      descriptor,
-      pageName: resolvedPageName,
-      rawSharedKey: resolvedRawSharedKey,
-      nodeId: resolvedNodeId,
-      expanded: expanded,
-    );
-  }
-
-  Future<void> _promptNewSharedModule(BuildContext context) async {
-    final controller = TextEditingController();
-    final result = await showDialog<String>(
-      context: context,
-      builder: (context) {
-        return AlertDialog(
-          title: const Text('Nouveau module partagé'),
-          content: TextField(
-            controller: controller,
-            autofocus: true,
-            decoration: const InputDecoration(
-              labelText: 'Nom du module',
-              hintText: 'ex: automatisations',
-            ),
-          ),
-          actions: [
-            TextButton(
-              onPressed: () => Navigator.of(context).pop(),
-              child: const Text('Annuler'),
-            ),
-            FilledButton(
-              onPressed: () {
-                Navigator.of(context).pop(controller.text.trim());
-              },
-              child: const Text('Créer'),
-            ),
-          ],
-        );
-      },
-    );
-    controller.dispose();
-    if (result == null) {
-      return;
-    }
-    final raw = result.trim();
-    if (raw.isEmpty) {
-      return;
-    }
-    final descriptor = ScriptDescriptor(
-      scope: ScriptScope.shared,
-      key: normaliseScriptKey(raw),
-    );
-    await _handleSelectScriptDescriptor(
-      descriptor,
-      rawSharedKey: raw,
-    );
-  }
-
-
-  String _normaliseCustomActionTemplate(String template) {
-    var value = template;
-    if (value.startsWith('\n')) {
-      value = value.substring(1);
-    }
-    if (!value.endsWith('\n')) {
-      value = '$value\n';
-    }
-    return value;
-  }
-
-  void _initialiseCustomActions() {
-    if (_customActions.isNotEmpty) {
-      return;
-    }
-    _customActions.addAll(<CustomAction>[
-      CustomAction(
-        id: 'log',
-        label: 'Ajouter un log',
-        template: _normaliseCustomActionTemplate('''
-  - log:
-      message: "Votre message"
-'''),
-      ),
-      CustomAction(
-        id: 'set_cell',
-        label: 'Ecrire une cellule',
-        template: _normaliseCustomActionTemplate('''
-  - set_cell:
-      cell: A1
-      value: "=B1"
-'''),
-      ),
-      CustomAction(
-        id: 'run_snippet',
-        label: 'Executer un snippet',
-        template: _normaliseCustomActionTemplate('''
-  - run_snippet:
-      module: shared/default
-      name: votre_snippet
-      args:
-        target: A1
-'''),
-      ),
-    ]);
-  }
-
-  Future<void> _handleSaveScript() async {
-    final activeTab = _activeScriptTab;
-    if (activeTab == null) {
-      setState(() {
-        _scriptEditorStatus =
-            "Impossible d'enregistrer: aucun script selectionne.";
-      });
-      return;
-    }
-    if (!activeTab.isMutable) {
-      setState(() {
-        _scriptEditorStatus =
-            'Edition indisponible sur cette plateforme (lecture seule).';
-      });
-      return;
-    }
-    final descriptor = _resolveScriptDescriptor();
-    if (descriptor == null) {
-      setState(() {
-        _scriptEditorStatus =
-            "Impossible d'enregistrer: aucun script selectionne.";
-      });
-      return;
-    }
-    setState(() {
-      _scriptEditorLoading = true;
-      _scriptEditorStatus = 'Enregistrement du script...';
-    });
-    try {
-      final stored =
-          await _runtime.storage.saveScript(descriptor, activeTab.controller.text);
-      await _runtime.reload();
-      await _refreshScriptLibrary(silent: true);
-      if (!mounted) {
-        return;
-      }
-      setState(() {
-        activeTab.descriptor = stored.descriptor;
-        activeTab.isDirty = false;
-        if (stored.descriptor.scope == ScriptScope.shared) {
-          activeTab.rawSharedKey = _sharedScriptKeyController.text;
-        }
-        _currentScriptDescriptor = stored.descriptor;
-        _scriptEditorStatus = 'Script enregistre dans ${stored.origin}.';
-        activeTab.status = _scriptEditorStatus;
-      });
-    } catch (error) {
-      if (!mounted) {
-        return;
-      }
-      setState(() {
-        final message =
-            "Erreur lors de l'enregistrement du script: $error";
-        _scriptEditorStatus = message;
-        activeTab.status = message;
-      });
-    } finally {
-      if (mounted) {
-        setState(() {
-          _scriptEditorLoading = false;
-        });
-      }
-    }
-  }
-
-  Future<void> _handleReloadScripts() async {
-    try {
-      setState(() {
-        _scriptEditorStatus = 'Rechargement des scripts...';
-      });
-      await _runtime.reload();
-      await _refreshScriptLibrary(silent: true);
-      final activeTab = _activeScriptTab;
-      if (activeTab != null) {
-        await _loadTabSource(activeTab);
-      }
-    } catch (error) {
-      if (!mounted) {
-        return;
-      }
-      setState(() {
-        _scriptEditorStatus = 'Erreur lors du rechargement: $error';
-      });
-    }
-  }
-
-  ScriptDescriptor? _resolveScriptDescriptor() {
-    switch (_scriptEditorScope) {
-      case ScriptScope.global:
-        return const ScriptDescriptor(scope: ScriptScope.global, key: 'default');
-      case ScriptScope.page:
-        final pageName = _scriptEditorPageName ??
-            (_manager.workbook.pages.isNotEmpty
-                ? _manager.workbook.pages.first.name
-                : null);
-        if (pageName == null) {
-          return null;
-        }
-        return ScriptDescriptor(
-          scope: ScriptScope.page,
-          key: normaliseScriptKey(pageName),
-        );
-      case ScriptScope.shared:
-        final raw = _sharedScriptKeyController.text.trim();
-        if (raw.isEmpty && _scriptSharedKey.isEmpty) {
-          return null;
-        }
-        final key = raw.isNotEmpty ? normaliseScriptKey(raw) : _scriptSharedKey;
-        return ScriptDescriptor(scope: ScriptScope.shared, key: key);
-    }
-  }
-
-  Future<void> _loadScriptEditor() async {
-    final descriptor = _resolveScriptDescriptor();
-    if (descriptor == null) {
-      setState(() {
-        _currentScriptDescriptor = null;
-        _scriptEditorLoading = false;
-        _scriptEditorMutable = false;
-        _scriptEditorStatus =
-            'Selectionnez un script a charger pour commencer.';
-      });
-      return;
-    }
-    final pageName =
-        descriptor.scope == ScriptScope.page ? _scriptEditorPageName : null;
-    final rawSharedKey = descriptor.scope == ScriptScope.shared
-        ? _sharedScriptKeyController.text
-        : null;
-    await _openOrFocusTab(
-      descriptor,
-      pageName: pageName,
-      rawSharedKey: rawSharedKey,
-    );
-  }
-
-  String _defaultScriptTemplate(ScriptDescriptor descriptor) {
-    switch (descriptor.scope) {
-      case ScriptScope.global:
-        return 'name: Global Script\n'
-            'scope: global\n'
-            'handlers:\n'
-            '  - event: workbook.open\n'
-            '    actions:\n'
-            '      - log:\n'
-            '          message: "Classeur ouvert"\n';
-      case ScriptScope.page:
-        return 'name: Page Script\n'
-            'scope: page\n'
-            'handlers:\n'
-            '  - event: page.enter\n'
-            '    actions:\n'
-            '      - log:\n'
-            '          message: "Bienvenue sur {{page.name}}"\n';
-      case ScriptScope.shared:
-        return 'name: Module partage\n'
-            'scope: shared\n'
-            'snippets:\n'
-            '  exemple:\n'
-            '    description: Exemple de snippet\n'
-            '    actions:\n'
-            '      - log:\n'
-            '          message: "Execution du snippet"\n';
-    }
-  }
-
-  Widget _buildAdminWorkspace(BuildContext context) {
-    final theme = Theme.of(context);
-    final workbook = _manager.workbook;
-    final pages = workbook.pages;
-    final isDark = theme.brightness == Brightness.dark;
-    final codeTheme = CodeThemeData(
-      styles: isDark ? monokaiSublimeTheme : githubTheme,
-    );
-    final lineNumberStyle = LineNumberStyle(
-      width: 48,
-      textStyle: theme.textTheme.bodySmall,
-    );
-    final descriptor = _currentScriptDescriptor;
-    final status = _scriptEditorStatus;
-    final scriptFileName = descriptor?.fileName;
-    final activeDescriptor = _descriptorForSelection() ?? descriptor;
-
-    final editorLayout = _buildAdminEditorLayout(
-      context: context,
-      codeTheme: codeTheme,
-      lineNumberStyle: lineNumberStyle,
-      pages: pages,
-      activeDescriptor: activeDescriptor,
-      scriptFileName: scriptFileName,
-      status: status,
-    );
-
-    final overlayBuilder = _scriptEditorOverlayBuilder ??
-        (_) => const SizedBox.shrink();
-
-    return Stack(
-      clipBehavior: Clip.none,
-      children: [
-        Card(
-          clipBehavior: Clip.antiAlias,
-          child: DefaultTabController(
-            length: 2,
-            child: Column(
-              crossAxisAlignment: CrossAxisAlignment.stretch,
-              children: [
-                Container(
-                  color: theme.colorScheme.surface,
-                  child: TabBar(
-                    labelColor: theme.colorScheme.primary,
-                    indicatorColor: theme.colorScheme.primary,
-                    tabs: const [
-                      Tab(icon: Icon(Icons.code), text: 'Scripts'),
-                      Tab(icon: Icon(Icons.menu_book_outlined), text: 'Documentation'),
-                    ],
-                  ),
-                ),
-                Expanded(
-                  child: TabBarView(
-                    children: [
-                      editorLayout,
-                      _buildAdminDocumentationTab(context),
-                    ],
-                  ),
-                ),
-              ],
-            ),
-          ),
-        ),
-        _ScriptEditorOverlayHost(
-          isActive: _scriptEditorFullscreen,
-          overlayBuilder: overlayBuilder,
-        ),
-      ],
-    );
-  }
-
-  void _handleExitScriptEditorFullscreen() {
-    if (!_scriptEditorFullscreen) {
-      return;
-    }
-    setState(() {
-      _scriptEditorFullscreen = false;
-    });
-  }
-
-  Widget _buildAdminEditorLayout({
-    required BuildContext context,
-    required CodeThemeData codeTheme,
-    required LineNumberStyle lineNumberStyle,
-    required List<WorkbookPage> pages,
-    required ScriptDescriptor? activeDescriptor,
-    required String? scriptFileName,
-    required String? status,
-  }) {
-    final theme = Theme.of(context);
-    final editorSurface = _buildScriptEditorSurface(
-      context: context,
-      codeTheme: codeTheme,
-      lineNumberStyle: lineNumberStyle,
-    );
-    final activeTab = _activeScriptTab;
-    final isDirty = activeTab?.isDirty ?? false;
-    final isMutable = activeTab?.isMutable ?? false;
-
-    Widget buildTabBar() {
-      if (_scriptEditorTabs.isEmpty) {
-        return Align(
-          alignment: Alignment.centerLeft,
-          child: Padding(
-            padding: const EdgeInsets.symmetric(vertical: 4),
-            child: Text(
-              'Aucun onglet ouvert. Sélectionnez un script dans la bibliothèque.',
-              style: theme.textTheme.bodySmall,
-            ),
-          ),
-        );
-      }
-
-      return SingleChildScrollView(
-        scrollDirection: Axis.horizontal,
-        child: Row(
-          children: [
-            for (var i = 0; i < _scriptEditorTabs.length; i++)
-              Padding(
-                padding: EdgeInsets.only(right: i == _scriptEditorTabs.length - 1 ? 0 : 8),
-                child: _buildScriptTabChip(
-                  context,
-                  tab: _scriptEditorTabs[i],
-                  isActive: i == _activeScriptTabIndex,
-                  onSelect: () => _activateTab(i),
-                  onClose: () => _handleCloseScriptTab(i),
-                ),
-              ),
-          ],
-        ),
-      );
-    }
-
-    List<Widget> buildActionButtons({required bool includeFullscreenToggle}) {
-      return [
-        IconButton(
-          tooltip:
-              _scriptEditorSplitPreview ? 'Fermer la vue scindée' : 'Afficher la vue scindée',
-          color: _scriptEditorSplitPreview ? theme.colorScheme.primary : null,
-          onPressed: () {
-            setState(() {
-              _scriptEditorSplitPreview = !_scriptEditorSplitPreview;
-            });
-          },
-          icon: const Icon(Icons.vertical_split),
-        ),
-        const SizedBox(width: 4),
-        if (includeFullscreenToggle) ...[
-          IconButton(
-            tooltip: _scriptEditorFullscreen
-                ? 'Quitter le plein écran'
-                : 'Afficher en plein écran',
-            color: _scriptEditorFullscreen ? theme.colorScheme.primary : null,
-            onPressed: () {
-              if (_scriptEditorFullscreen) {
-                _handleExitScriptEditorFullscreen();
-              } else {
-                setState(() {
-                  _scriptEditorFullscreen = true;
-                });
-              }
-            },
-            icon: Icon(
-              _scriptEditorFullscreen
-                  ? Icons.close_fullscreen
-                  : Icons.open_in_full,
-            ),
-          ),
-          const SizedBox(width: 4),
-        ],
-        IconButton(
-          tooltip: 'Recharger tous les scripts',
-          onPressed: _scriptEditorLoading ? null : _handleReloadScripts,
-          icon: const Icon(Icons.refresh),
-        ),
-        const SizedBox(width: 4),
-        FilledButton.icon(
-          onPressed: (_scriptEditorLoading || !isDirty || !isMutable)
-              ? null
-              : _handleSaveScript,
-          icon: const Icon(Icons.save_outlined),
-          label: Text(
-            isDirty ? 'Enregistrer*' : 'Enregistrer',
-          ),
-        ),
-      ];
-    }
-
-    Widget buildEditorContent({required bool fullscreen}) {
-      return Column(
-        crossAxisAlignment: CrossAxisAlignment.stretch,
-        children: [
-          buildTabBar(),
-          if (_scriptEditorTabs.isNotEmpty) const SizedBox(height: 8),
-          if (scriptFileName != null)
-            Text(
-              'Fichier actuel : $scriptFileName',
-              style: theme.textTheme.bodySmall,
-            ),
-          if (scriptFileName != null) const SizedBox(height: 8),
-          if (_customActions.isNotEmpty) _buildCustomActionsBar(context),
-          if (_customActions.isNotEmpty) const SizedBox(height: 12),
-          if (fullscreen)
-            Expanded(child: editorSurface)
-          else
-            Flexible(fit: FlexFit.tight, child: editorSurface),
-          const SizedBox(height: 8),
-          if (status != null)
-            Text(
-              status,
-              style: theme.textTheme.bodySmall,
-            ),
-        ],
-      );
-    }
-
-    final baseLayout = Column(
-      crossAxisAlignment: CrossAxisAlignment.stretch,
-      children: [
-        Padding(
-          padding: const EdgeInsets.fromLTRB(16, 16, 16, 12),
-          child: Row(
-            children: [
-              Expanded(
-                child: Text(
-                  'Espace de développement',
-                  style: theme.textTheme.titleMedium,
-                ),
-              ),
-              ...buildActionButtons(includeFullscreenToggle: true),
-            ],
-          ),
-        ),
-        const Divider(height: 1),
-        Expanded(
-          child: Row(
-            crossAxisAlignment: CrossAxisAlignment.stretch,
-            children: [
-              if (!_scriptEditorFullscreen) ...[
-                SizedBox(
-                  width: 240,
-                  child: _buildScriptLibraryPanel(context),
-                ),
-                const VerticalDivider(width: 1),
-              ],
-              Expanded(
-                child: Padding(
-                  padding: const EdgeInsets.fromLTRB(16, 16, 16, 12),
-                  child: buildEditorContent(fullscreen: _scriptEditorFullscreen),
-                ),
-              ),
-            ],
-          ),
-        ),
-      ],
-    );
-
-    _scriptEditorOverlayBuilder = (_) {
-      return Positioned.fill(
-        child: Theme(
-          data: theme,
-          child: Scaffold(
-            backgroundColor: theme.colorScheme.surface,
-            body: SafeArea(
-              child: Column(
-                crossAxisAlignment: CrossAxisAlignment.stretch,
-                children: [
-                  Padding(
-                    padding: const EdgeInsets.fromLTRB(16, 16, 16, 12),
-                    child: Row(
-                      children: [
-                        Expanded(
-                          child: Text(
-                            'Espace de développement',
-                            style: theme.textTheme.titleMedium,
-                          ),
-                        ),
-                        FilledButton.icon(
-                          onPressed: _handleExitScriptEditorFullscreen,
-                          icon: const Icon(Icons.fullscreen_exit),
-                          label: const Text('Quitter le plein écran'),
-                        ),
-                        const SizedBox(width: 12),
-                        ...buildActionButtons(includeFullscreenToggle: false),
-                      ],
-                    ),
-                  ),
-                  const Divider(height: 1),
-                  Expanded(
-                    child: Row(
-                      crossAxisAlignment: CrossAxisAlignment.stretch,
-                      children: [
-                        SizedBox(
-                          width: 240,
-                          child: _buildScriptLibraryPanel(context),
-                        ),
-                        const VerticalDivider(width: 1),
-                        Expanded(
-                          child: Padding(
-                            padding: const EdgeInsets.fromLTRB(16, 16, 16, 12),
-                            child: buildEditorContent(fullscreen: true),
-                          ),
-                        ),
-                      ],
-                    ),
-                  ),
-                ],
-              ),
-            ),
-          ),
-        ),
-      );
-    };
-
-    return baseLayout;
-  }
-
-  Widget _buildScriptEditorSurface({
-    required BuildContext context,
-    required CodeThemeData codeTheme,
-    required LineNumberStyle lineNumberStyle,
-  }) {
-    final theme = Theme.of(context);
-    final borderDecoration = BoxDecoration(
-      border: Border.all(
-        color: theme.colorScheme.outline.withOpacity(0.25),
-      ),
-      borderRadius: const BorderRadius.all(Radius.circular(8)),
-    );
-
-    final activeTab = _activeScriptTab;
-    final controller = activeTab?.controller;
-    final isMutable = activeTab?.isMutable ?? false;
-
-    Widget buildEditorField() {
-      if (controller == null) {
-        return DecoratedBox(
-          decoration: borderDecoration,
-          child: Center(
-            child: Padding(
-              padding: const EdgeInsets.all(24),
-              child: Text(
-                'Ouvrez un script depuis la bibliothèque pour commencer.',
-                style: theme.textTheme.bodyMedium,
-                textAlign: TextAlign.center,
-              ),
-            ),
-          ),
-        );
-      }
-
-      final field = CodeTheme(
-        data: codeTheme,
-        child: DecoratedBox(
-          decoration: borderDecoration,
-          child: Stack(
-            children: [
-              Positioned.fill(
-                child: IgnorePointer(
-                  ignoring: !isMutable,
-                  child: TopAlignedCodeField(
-                    controller: controller,
-                    expands: true,
-                    textStyle: const TextStyle(
-                      fontFamily: 'monospace',
-                      fontSize: 13,
-                    ),
-                    lineNumberStyle: lineNumberStyle,
-                    padding: const EdgeInsets.all(12),
-                    background: theme.colorScheme.surface,
-                    textAlignVertical: TextAlignVertical.top,
-                  ),
-                ),
-              ),
-              if (_scriptEditorLoading)
-                const Positioned(
-                  top: 16,
-                  right: 16,
-                  child: SizedBox(
-                    width: 18,
-                    height: 18,
-                    child: CircularProgressIndicator(strokeWidth: 2),
-                  ),
-                ),
-            ],
-          ),
-        ),
-      );
-
-      return field;
-    }
-
-    final editor = buildEditorField();
-
-    Widget buildPreview() {
-      return DecoratedBox(
-        decoration: borderDecoration,
-        child: Material(
-          type: MaterialType.transparency,
-          child: SingleChildScrollView(
-            padding: const EdgeInsets.all(12),
-            child: SelectableText(
-              controller == null || controller.text.isEmpty
-                  ? 'Aucun contenu pour le moment.'
-                  : controller.text,
-              style: const TextStyle(
-                fontFamily: 'monospace',
-                fontSize: 13,
-              ),
-            ),
-          ),
-        ),
-      );
-    }
-
-    return LayoutBuilder(
-      builder: (context, constraints) {
-        Widget content = editor;
-
-        if (_scriptEditorSplitPreview) {
-          content = SizedBox(
-            height: constraints.maxHeight,
-            child: Row(
-              crossAxisAlignment: CrossAxisAlignment.stretch,
-              children: [
-                Expanded(child: editor),
-                const VerticalDivider(width: 1),
-                Expanded(child: buildPreview()),
-              ],
-            ),
-          );
-        }
-
-        return AnimatedSize(
-          duration: const Duration(milliseconds: 200),
-          curve: Curves.easeInOut,
-          alignment: Alignment.topCenter,
-          child: SizedBox(
-            width: constraints.maxWidth,
-            height: constraints.maxHeight,
-            child: content,
-          ),
-        );
-      },
-    );
-  }
-
-
-  Widget _buildScriptLibraryPanel(BuildContext context) {
-    final theme = Theme.of(context);
-    final colorScheme = theme.colorScheme;
-    final canToggle = _scriptTreeExpandableNodes.isNotEmpty;
-    final hasCollapsed = _scriptTreeExpandableNodes
-        .any((id) => !(_scriptTreeExpanded[id] ?? true));
-    final toggleLabel = hasCollapsed ? 'Tout déplier' : 'Tout replier';
-
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.stretch,
-      children: [
-        Padding(
-          padding: const EdgeInsets.fromLTRB(16, 16, 16, 4),
-          child: Row(
-            children: [
-              Expanded(
-                child: Text(
-                  'Bibliothèque de scripts',
-                  style: theme.textTheme.titleSmall,
-                ),
-              ),
-              if (canToggle)
-                TextButton.icon(
-                  onPressed: _toggleScriptTreeExpansion,
-                  icon: Icon(hasCollapsed ? Icons.unfold_more : Icons.unfold_less),
-                  label: Text(toggleLabel),
-                  style: TextButton.styleFrom(
-                    padding: const EdgeInsets.symmetric(horizontal: 12, vertical: 8),
-                  ),
-                ),
-            ],
-          ),
-        ),
-        Expanded(
-          child: _scriptLibraryLoading
-              ? const Center(child: CircularProgressIndicator(strokeWidth: 2))
-              : ListView.separated(
-                  padding: const EdgeInsets.only(bottom: 12),
-                  itemCount: _scriptTreeNodes.length + 1,
-                  separatorBuilder: (context, index) => const SizedBox(height: 8),
-                  itemBuilder: (context, index) {
-                    if (index == _scriptTreeNodes.length) {
-                      return Padding(
-                        padding: const EdgeInsets.fromLTRB(16, 12, 16, 4),
-                        child: OutlinedButton.icon(
-                          onPressed: () => _promptNewSharedModule(context),
-                          icon: const Icon(Icons.add),
-                          label: const Text('Nouveau module partagé'),
-                        ),
-                      );
-                    }
-                    final node = _scriptTreeNodes[index];
-                    return _buildScriptTreeNode(context, node);
-                  },
-                ),
-        ),
-        if (_scriptLibraryError != null)
-          Padding(
-            padding: const EdgeInsets.fromLTRB(16, 0, 16, 12),
-            child: Text(
-              _scriptLibraryError!,
-              style: theme.textTheme.bodySmall?.copyWith(
-                color: colorScheme.error,
-              ),
-            ),
-          ),
-      ],
-    );
-  }
-
-  Widget _buildAdminDocumentationTab(BuildContext context) {
-    final theme = Theme.of(context);
-    final isDark = theme.brightness == Brightness.dark;
-    final codeBackground = isDark
-        ? theme.colorScheme.surfaceVariant.withOpacity(0.4)
-        : theme.colorScheme.surfaceVariant;
-
-    return ListView(
-      padding: const EdgeInsets.all(16),
-      children: [
-        Text(
-          'Guide de référence rapide',
-          style: theme.textTheme.titleMedium,
-        ),
-        const SizedBox(height: 12),
-        Text(
-          'Les scripts Optima sont écrits en YAML. Chaque script définit un nom, une portée (global, page ou module partagé) et une liste de gestionnaires d’évènements.',
-          style: theme.textTheme.bodyMedium,
-        ),
-        const SizedBox(height: 16),
-        Text('Événements disponibles', style: theme.textTheme.titleSmall),
-        const SizedBox(height: 8),
-        _buildDocBullet(
-          context,
-          'workbook.open',
-          'Déclenché lors de l’ouverture du classeur.',
-        ),
-        _buildDocBullet(
-          context,
-          'workbook.close',
-          'Déclenché à la fermeture du classeur.',
-        ),
-        _buildDocBullet(
-          context,
-          'page.enter',
-          'Appelé quand un utilisateur arrive sur une page.',
-        ),
-        _buildDocBullet(
-          context,
-          'page.leave',
-          'Appelé avant de quitter la page active.',
-        ),
-        _buildDocBullet(
-          context,
-          'cell.changed',
-          'Notifié lorsqu’une cellule change de valeur.',
-        ),
-        _buildDocBullet(
-          context,
-          'selection.changed',
-          'Notifié lorsqu’une sélection de cellules est modifiée.',
-        ),
-        _buildDocBullet(
-          context,
-          'notes.changed',
-          'Déclenché lorsque le contenu d’une page de notes est édité.',
-        ),
-        const SizedBox(height: 16),
-        Text('Actions supportées', style: theme.textTheme.titleSmall),
-        const SizedBox(height: 8),
-        _buildDocBullet(
-          context,
-          'log',
-          'Affiche un message dans la console des scripts. Utilisez le paramètre "message" pour personnaliser le texte.',
-        ),
-        _buildDocBullet(
-          context,
-          'set_cell',
-          'Écrit une valeur dans une cellule (paramètres : cell, sheet?, value/raw). Les expressions sont évaluées après substitution des variables de contexte.',
-        ),
-        _buildDocBullet(
-          context,
-          'clear_cell',
-          'Efface le contenu d’une cellule ciblée.',
-        ),
-        _buildDocBullet(
-          context,
-          'run_snippet',
-          'Exécute un snippet défini dans un module partagé (paramètres : module, name, args?).',
-        ),
-        const SizedBox(height: 16),
-        Text('Contexte disponible', style: theme.textTheme.titleSmall),
-        const SizedBox(height: 8),
-        Text(
-          'Les templates peuvent accéder aux informations du classeur : {{workbook.pageCount}}, {{page.name}}, {{sheetKey}}… Utilisez ces variables pour créer des scripts dynamiques.',
-          style: theme.textTheme.bodyMedium,
-        ),
-        const SizedBox(height: 16),
-        Text('Exemple complet', style: theme.textTheme.titleSmall),
-        const SizedBox(height: 8),
-        DecoratedBox(
-          decoration: BoxDecoration(
-            color: codeBackground,
-            borderRadius: const BorderRadius.all(Radius.circular(8)),
-            border: Border.all(
-              color: theme.colorScheme.outline.withOpacity(0.4),
-            ),
-          ),
-          child: const Padding(
-            padding: EdgeInsets.all(12),
-            child: SelectableText(
-              'name: Exemple page\nscope: page\nhandlers:\n  - event: page.enter\n    actions:\n      - log:\n          message: "Bienvenue {{page.name}}"\n      - set_cell:\n          cell: A1\n          value: "=SUM(B1:B5)"\n',
-              style: TextStyle(
-                fontFamily: 'monospace',
-                height: 1.4,
-              ),
-            ),
-          ),
-        ),
-        const SizedBox(height: 16),
-        Text(
-          'Astuce : utilisez la bibliothèque de pré-code pour insérer un squelette d’actions avant de personnaliser votre script.',
-          style: theme.textTheme.bodySmall,
-        ),
-      ],
-    );
-  }
-
-  Widget _buildDocBullet(
-    BuildContext context,
-    String title,
-    String description,
-  ) {
-    final theme = Theme.of(context);
-    return Padding(
-      padding: const EdgeInsets.symmetric(vertical: 4),
-      child: Row(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          const Text('• '),
-          Expanded(
-            child: RichText(
-              text: TextSpan(
-                style: theme.textTheme.bodyMedium,
-                children: [
-                  TextSpan(
-                    text: '$title : ',
-                    style: theme.textTheme.bodyMedium?.copyWith(
-                      fontWeight: FontWeight.w600,
-                    ),
-                  ),
-                  TextSpan(text: description),
-                ],
-              ),
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-
-  Widget _buildCustomActionsBar(BuildContext context) {
-    final theme = Theme.of(context);
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        Text(
-          'Pré-code',
-          style: theme.textTheme.titleSmall,
-        ),
-        const SizedBox(height: 8),
-        Wrap(
-          spacing: 8,
-          runSpacing: 8,
-          children: _customActions
-              .map(
-                (action) => Tooltip(
-                  message: action.template,
-                  preferBelow: false,
-                  child: ActionChip(
-                    label: Text(action.label),
-                    onPressed: () => _handleInsertCustomAction(action),
-                  ),
-                ),
-              )
-              .toList(growable: false),
-        ),
-      ],
-    );
-  }
-
-  void _handleInsertCustomAction(CustomAction action) {
-    final tab = _activeScriptTab;
-    if (tab == null) {
-      return;
-    }
-    final controller = tab.controller;
-    final selection = controller.selection;
-    final insertion = action.template;
-    final text = controller.text;
-    final start = selection.isValid ? selection.start : text.length;
-    final end = selection.isValid ? selection.end : text.length;
-    final newText = text.replaceRange(start, end, insertion);
-    final newSelection = TextSelection.collapsed(
-      offset: start + insertion.length,
-    );
-    _suppressScriptEditorChanges = true;
-    controller.value = controller.value.copyWith(
-      text: newText,
-      selection: newSelection,
-      composing: TextRange.empty,
-    );
-    _suppressScriptEditorChanges = false;
-    if (!tab.isDirty || _scriptEditorSplitPreview) {
-      setState(() {
-        tab.isDirty = true;
-      });
-    } else {
-      tab.isDirty = true;
-    }
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return AnimatedBuilder(
-      animation: _manager,
-      builder: (context, _) {
-        final workbook = _manager.workbook;
-        final pages = workbook.pages;
-        final activePageIndex = _manager.activePageIndex;
-
-        final tabs = <WorkbookPageTabData>[];
-        for (var pageIndex = 0; pageIndex < pages.length; pageIndex++) {
-          final page = pages[pageIndex];
-          int? sheetIndex;
-          if (page is Sheet) {
-            final index = workbook.sheets.indexOf(page);
-            if (index != -1) {
-              sheetIndex = index;
-            }
-          }
-          tabs.add(
-            WorkbookPageTabData(
-              title: page.name,
-              pageIndex: pageIndex,
-              icon: workbookPageIcon(page),
-              sheetIndex: sheetIndex,
-              canClose: page is Sheet && workbook.sheets.length > 1,
-            ),
-          );
-        }
-
-        final theme = Theme.of(context);
-        final borderColor = theme.colorScheme.outline.withOpacity(0.15);
-
-        final workbookColumn = Column(
-          crossAxisAlignment: CrossAxisAlignment.stretch,
-          children: [
-            CommandRibbon(
-              commandManager: _manager,
-              onBeforeCommand: _commitActiveSelectionEdits,
-            ),
-            WorkbookPageTabBar(
-              tabs: tabs,
-              selectedPageIndex: activePageIndex,
-              onSelectPage: _handleSelectPage,
-            ),
-            Expanded(
-              child: pages.isEmpty
-                  ? const Center(child: Text('Aucune page disponible'))
-                  : PageView.builder(
-                      controller: _pageController,
-                      itemCount: pages.length,
-                      onPageChanged: (index) {
-
-                        final currentWorkbook = _manager.workbook;
-
-                        if (_currentPageIndex >= 0 &&
-
-                            _currentPageIndex < currentWorkbook.pages.length) {
-
-                          final previousPage =
-
-                              currentWorkbook.pages[_currentPageIndex];
-
-                          unawaited(_runtime.dispatchPageLeave(previousPage));
-
-                        }
-
-                        _commitEditsForPage(currentWorkbook, _currentPageIndex);
-
-                        _currentPageIndex = index;
-
-                        if (index >= 0 && index < currentWorkbook.pages.length) {
-
-                          final nextPage = currentWorkbook.pages[index];
-
-                          unawaited(_runtime.ensurePageScript(nextPage));
-
-                          unawaited(_runtime.dispatchPageEnter(nextPage));
-
-                          if (_isAdmin &&
-                              _scriptEditorScope == ScriptScope.page) {
-                            _scriptEditorPageName = nextPage.name;
-                            unawaited(_loadScriptEditor());
-                          }
-
-                        }
-
-                        _manager.setActivePage(index);
-
-                      },
-
-                      itemBuilder: (context, index) {
-                        final page = pages[index];
-                        if (page is Sheet) {
-                          final selectionState =
-                              _stateForSheet(workbook, page);
-                          return Padding(
-                            padding: const EdgeInsets.all(16),
-                            child: Column(
-                              crossAxisAlignment: CrossAxisAlignment.stretch,
-                              children: [
-                                FormulaBar(selectionState: selectionState),
-                                const SizedBox(height: 16),
-                                Expanded(
-                                  child: DecoratedBox(
-                                    decoration: BoxDecoration(
-                                      borderRadius: BorderRadius.circular(12),
-                                      color: Theme.of(context)
-                                          .colorScheme
-                                          .surface,
-                                      border: Border.all(
-                                        color: Theme.of(context)
-                                            .dividerColor
-                                            .withOpacity(0.4),
-                                      ),
-                                    ),
-                                    child: ClipRRect(
-                                      borderRadius: BorderRadius.circular(12),
-                                      child: SheetGrid(
-                                        selectionState: selectionState,
-                                        rowCount: page.rowCount,
-                                        columnCount: page.columnCount,
-                                      ),
-                                    ),
-                                  ),
-                                ),
-                              ],
-                            ),
-                          );
-                        }
-                        if (page is MenuPage) {
-                          return MenuPageView(
-
-                            page: page,
-
-                            workbook: workbook,
-
-                            onOpenPage: _handleSelectPage,
-
-                            onCreateSheet: _handleAddSheet,
-
-                            onCreateNotes: _handleAddNotesPage,
-
-                            onRemovePage: _handleRemovePage,
-
-                            canRemovePage: (index) => _canRemovePage(workbook, index),
-
-                            enableEditing: _isAdmin,
-
-                          );
-
-                        }
-                        if (page is NotesPage) {
-                          final controller =
-                              _controllerForNotesPage(page);
-                          return NotesPageView(
-                            page: page,
-                            controller: controller,
-                          );
-                        }
-                        return Center(
-                          child: Text('Page inconnue : ${page.name}'),
-                        );
-                      },
-                    ),
-                  ),
-          ],
-        );
-
-        final workbookSurface = Container(
-          decoration: BoxDecoration(
-            color: theme.colorScheme.surface,
-            borderRadius: BorderRadius.circular(12),
-            border: Border.all(color: borderColor),
-          ),
-          child: ClipRRect(
-            borderRadius: BorderRadius.circular(12),
-            child: workbookColumn,
-          ),
-        );
-
-        if (_isAdmin) {
-          return LayoutBuilder(
-            builder: (context, constraints) {
-              final isWide = constraints.maxWidth >= 1100;
-              if (isWide) {
-                return Row(
-                  crossAxisAlignment: CrossAxisAlignment.stretch,
-                  children: [
-                    Expanded(
-                      flex: 7,
-                      child: Padding(
-                        padding: const EdgeInsets.fromLTRB(16, 16, 12, 16),
-                        child: workbookSurface,
-                      ),
-                    ),
-                    const VerticalDivider(width: 1),
-                    Expanded(
-                      flex: 5,
-                      child: Padding(
-                        padding: const EdgeInsets.fromLTRB(12, 16, 16, 16),
-                        child: _buildAdminWorkspace(context),
-                      ),
-                    ),
-                  ],
-                );
-              }
-              return Column(
-                children: [
-                  Expanded(
-                    child: Padding(
-                      padding: const EdgeInsets.fromLTRB(16, 16, 16, 8),
-                      child: workbookSurface,
-                    ),
-                  ),
-                  const Divider(height: 1),
-                  SizedBox(
-                    height: 420,
-                    child: Padding(
-                      padding: const EdgeInsets.fromLTRB(16, 12, 16, 16),
-                      child: _buildAdminWorkspace(context),
-                    ),
-                  ),
-                ],
-              );
-            },
-          );
-        }
-
-        return Padding(
-          padding: const EdgeInsets.all(16),
-          child: workbookSurface,
-        );
-      },
-    );
-  }
-}
-
-class _ScriptEditorOverlayHost extends StatefulWidget {
-  const _ScriptEditorOverlayHost({
-    required this.isActive,
-    required this.overlayBuilder,
-  });
-
-  final bool isActive;
-  final WidgetBuilder overlayBuilder;
-
-  @override
-  State<_ScriptEditorOverlayHost> createState() =>
-      _ScriptEditorOverlayHostState();
-}
-
-class _ScriptEditorOverlayHostState extends State<_ScriptEditorOverlayHost> {
-  OverlayEntry? _entry;
-  bool _overlayRebuildScheduled = false;
-
-  void _scheduleOverlayRebuild() {
-    if (_overlayRebuildScheduled) {
-      return;
-    }
-    _overlayRebuildScheduled = true;
-    WidgetsBinding.instance.addPostFrameCallback((_) {
-      _overlayRebuildScheduled = false;
-      if (!mounted) {
-        return;
-      }
-      _entry?.markNeedsBuild();
-    });
-  }
-
-  @override
-  void initState() {
-    super.initState();
-    if (widget.isActive) {
-      _insertEntry();
-    }
-  }
-
-  @override
-  void didUpdateWidget(covariant _ScriptEditorOverlayHost oldWidget) {
-    super.didUpdateWidget(oldWidget);
-    if (!oldWidget.isActive && widget.isActive) {
-      _insertEntry();
-    } else if (oldWidget.isActive && !widget.isActive) {
-      _removeEntry();
-    } else if (widget.isActive && _entry != null) {
-      _scheduleOverlayRebuild();
-    }
-  }
-
-  void _insertEntry() {
-    WidgetsBinding.instance.addPostFrameCallback((_) {
-      if (!mounted || !widget.isActive) {
-        return;
-      }
-      if (_entry != null) {
-        _scheduleOverlayRebuild();
-        return;
-      }
-      final overlay = Overlay.of(context, rootOverlay: true);
-      if (overlay == null) {
-        return;
-      }
-      _entry = OverlayEntry(
-        builder: (context) => widget.overlayBuilder(context),
-      );
-      overlay.insert(_entry!);
-    });
-  }
-
-  void _removeEntry() {
-    _entry?.remove();
-    _entry = null;
-  }
-
-  @override
-  void dispose() {
-    _removeEntry();
-    super.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    if (widget.isActive && _entry != null) {
-      _scheduleOverlayRebuild();
-    }
-    return const SizedBox.shrink();
-  }
-}
-
